+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Use a font sizes larger or as large as 15px
 - Change not reservable colour into blue and selected colour into white
 - Don't show availability status for machine room
+- Recording studio button opens tooltip on top of the rightmost room in the recording studio setup
 
 ### Fixed
 - Fix untranslated "you are here" label
@@ -22,7 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fix room data for retro Gaming corner, Gaming hall and Kuutio
 - Fix translations coverstitch machine, printing room and language menu
 - Fix naming of Studio 1
-- FIx game room order
+- Fix game room order
+- Fix you are here label overlapping position indicator in English translation
 
 ## [0.1.0] - 2020-04-16
 ### Added